update: fixed photogallery section and added pure react carousel
