+++ packages/visual-editor/package.json
@@ -60,6 +60,7 @@
     "picocolors": "^1.0.1",
     "prettier": "^3.3.3",
     "prompts": "^2.4.2",
+    "pure-react-carousel": "^1.32.0",
     "react-color": "^2.19.3",
     "react-error-boundary": "^5.0.0",
     "sonner": "^1.5.0",
